+++ BlazorLib/FluentNavMenu/NavMenu.razor
@@ -1,4 +1,4 @@
-﻿@namespace BlazorSite.Shared
+﻿@namespace BlazorLib.FluentNavMenu
 @inherits FluentComponentBase
 <div>
     <CascadingValue Value="this" Name="NavMenu" IsFixed="true">
