+++ BlazorLib/FluentNavMenu/NavMenu.razor.cs
@@ -2,7 +2,7 @@
 using Microsoft.Fast.Components.FluentUI;
 using Microsoft.Fast.Components.FluentUI.Utilities;
 
-namespace BlazorSite.Shared;
+namespace BlazorLib.FluentNavMenu;
 
 public partial class NavMenu : FluentComponentBase
 {
