+++ BlazorLib/FluentNavMenu/NavMenuGroup.razor
@@ -1,5 +1,5 @@
 ﻿
-@namespace BlazorSite.Shared
+@namespace BlazorLib.FluentNavMenu
 @inherits FluentComponentBase
 <FluentTreeItem Id="@Id"
                 Class="@ClassValue"
