+++ BlazorLib/FluentNavMenu/NavMenuLink.razor
@@ -1,4 +1,4 @@
-﻿@namespace BlazorSite.Shared
+﻿@namespace BlazorLib.FluentNavMenu
 @inherits FluentComponentBase
 <FluentTreeItem Id="@Id"
                 Class="@ClassValue"
