+++ BlazorLib/_Imports.razor
@@ -1,2 +1,3 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.Fast.Components.FluentUI
