+++ BlazorSite/Shared/MainLayout.razor
@@ -1,5 +1,5 @@
-﻿@namespace BlazorSite.Shared
-@inherits LayoutComponentBase
+﻿@inherits LayoutComponentBase
+@inject NavigationManager Navigation
 
 
     
@@ -7,13 +7,13 @@
     <div class="topbar">
         <div class="main-menu">
             <NavMenu>
-                <NavMenuLink Text="PromiPay " @onclick="OnClick" />
-                <NavMenuLink Text="PromiMart " @onclick="OnClick" />
-                <NavMenuLink Text="Settings " @onclick="OnClick" />
-                <NavMenuLink Text="Help " @onclick="OnClick" />
-                <NavMenuLink Text="Our Team " @onclick="OnClick" />
-                <NavMenuLink Text="Our Misson " @onclick="OnClick" />
-                <NavMenuLink Text="Join Us! " @onclick="OnClick" />
+                <NavMenuLink Text="PromiPay" @onclick="OnClickPay" />
+                <NavMenuLink Text="PromiMart" @onclick="OnClickMart" />
+                <NavMenuLink Text="Settings" @onclick="OnClick" />
+                <NavMenuLink Text="Help" @onclick="OnClick" />
+                <NavMenuLink Text="Our Team" @onclick="OnClick" />
+                <NavMenuLink Text="Our Misson" @onclick="OnClick" />
+                <NavMenuLink Text="Join Us!" @onclick="OnClick" />
             </NavMenu>
         </div>
         <div class="brand">
@@ -39,6 +39,16 @@
     void OnClick(MouseEventArgs e)
     {
         
+    }
+
+    void OnClickPay(MouseEventArgs e)
+    {
+        
+    }
+
+    void OnClickMart(MouseEventArgs e)
+    {
+        Navigation.NavigateTo("/mart");
     }
 
     protected override void OnAfterRender(bool firstRender) {
