+++ BlazorSite/_Imports.razor
@@ -12,7 +12,8 @@
 @using PromiseLib
 @using PromiseLib.Weather
 @using BlazorLib
+@using BlazorLib.FluentNavMenu
 @using Microsoft.Fast.Components.FluentUI
-@using Microsoft.Fast.Components.FluentUI.DesignTokens;
+@using Microsoft.Fast.Components.FluentUI.DesignTokens
